* Source/NSBitmapImageRep+ICNS.m: Redefine icns_size_t to int, as   this needs to be a 32 bit integer and on 64 bit systems long   uses 64 bits.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@28501 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSBitmapImageRep+ICNS.m
@@ -55,7 +55,8 @@
  */
 
 typedef unsigned char icns_byte_t;
-typedef unsigned long icns_size_t;
+// must be a 32 bit integer
+typedef unsigned int icns_size_t;
 typedef struct _icns_type_t {
   char c[4];
 } icns_type_t;
